refactor(ml): Migrate to typescript
