+++ packages/ml/package.json
@@ -3,12 +3,13 @@
   "version": "23.8.6",
   "author": "Invertase <oss@invertase.io> (http://invertase.io)",
   "description": "React Native Firebase - Firebase ML brings the power of machine learning vision to your React Native application, supporting both Android & iOS.",
-  "main": "lib/index.js",
-  "types": "lib/index.d.ts",
+  "main": "./dist/module/index.js",
+  "types": "./dist/typescript/lib/index.d.ts",
   "scripts": {
-    "build": "genversion --semi lib/version.js",
+    "build": "genversion --esm --semi lib/version.ts",
     "build:clean": "rimraf android/build && rimraf ios/build",
-    "prepare": "yarn run build"
+    "compile": "bob build",
+    "prepare": "yarn run build && yarn compile"
   },
   "repository": {
     "type": "git",
@@ -31,5 +32,26 @@
   "publishConfig": {
     "access": "public",
     "provenance": true
-  }
+  },
+  "devDependencies": {
+    "react-native-builder-bob": "^0.40.17",
+    "typescript": "^5.9.3"
+  },
+  "exports": {
+    ".": {
+      "source": "./lib/index.ts",
+      "types": "./dist/typescript/lib/index.d.ts",
+      "default": "./dist/module/index.js"
+    },
+    "./package.json": "./package.json"
+  },
+  "react-native-builder-bob": {
+    "source": "lib",
+    "output": "dist",
+    "targets": [
+      ["module", { "esm": true }],
+      ["typescript", { "tsc": "../../node_modules/.bin/tsc" }]
+    ]
+  },
+  "eslintIgnore": ["node_modules/", "dist/"]
 }
