+++ packages/ml/tsconfig.json
@@ -1,0 +1,15 @@
+{
+  "extends": "../../tsconfig.packages.base.json",
+  "compilerOptions": {
+    "baseUrl": ".",
+    "rootDir": ".",
+    "moduleResolution": "node",
+    "paths": {
+      "@react-native-firebase/app": ["../app/dist/typescript/lib"],
+      "@react-native-firebase/app/dist/module/common": ["../app/dist/typescript/lib/common"],
+      "@react-native-firebase/app/dist/module/internal": ["../app/dist/typescript/lib/internal"]
+    }
+  },
+  "include": ["lib/**/*"],
+  "exclude": ["node_modules", "dist", "__tests__", "**/*.test.ts"]
+}
